Исправлен  метод удаления избранной вакансии

+++ app/src/main/java/ru/practicum/android/diploma/db/data/db/entity/AreaEntity.kt
@@ -7,6 +7,6 @@ import androidx.room.PrimaryKey
 data class AreaEntity(
     @PrimaryKey
     val id: String,
-    val employerId: Int,
+    val employerId: Long,
     val name: String
 )
